Refactor secure_rsa_db module to add new secure_rsa row
This commit refactors the secure_rsa_db module to add a new function `add_secure_rsa` that adds a new row to the `secure_rsa` table. The function also sets the previous valid row to invalid. This change improves the functionality of the module and allows for better management of secure_rsa values.